backend/requirements.txt
Flask
Flask-Cors
pytest
Flask-SQLAlchemy
SQLAlchemy
openai
python-dotenv 